Add files via upload
I am dealing with a binary classification (0 or 1 & yes or no). The range of the dataset is 2,400*76 columns. In this dataset, all values are numerical (float or int) values. All values are on the balance sheet, P&L, and ratios. This is a supervised learning problem. There are no null values in this dataset. I am trying to predict whether they will default or not. Because I am predicting the company has to be capable of repaying the loan. So, I can begin "normalizing" here with my first step. From this data, 93% of companies' non-default means this company is capable of repaying the loan and remains 7% of companies likely to default. So, banks will most likely reject these companies. For this total
of 76 columns, I created a correlation heat map with 76 column values. The correlation map is not clear and it’s not possible to make any assumptions. So, Idecided to take only ratios, and in our model, ratios are very important. After considering only ratios, there are a total of 29 columns out of 76. Again, I created a correlation map with 29 variables. I found some of the values are highly correlated with each other, like PAT, PBT, current ratio, and liquidity ratio. I took only the important values from these highly correlated values. For example, the current ratio and liquidity ratio are almost the same, so I chose the current ratio. After selecting particular ratios and removing a few of the ratios, I got a total of 22 columns. So, in my next step, I started running logistic regression. In logistic regression, I added a constant, the constant is used to prevent overall bias by forcing the equal to zero. In my training feature, the target value is set to default. Then I separated the dependent and independent
variables. Then I run the logistic regression. After running the code, I removed the default. Then I start checking the significance of the variables based on my p-value My significance level is rejected if my p-value is greater than 0.05, and if my p-value is less than 0.05, my significance level is accepted. I repeated the logistic regression with four variables again. This time my p-value is 0, which means my null hypothesis is rejected and my test has statistical significance. Four significant ratios are identified after comparing them to the P-value. These four ratios are good enough to predict the company's likely default or nondefault [1) PAT 2) TAX LIABILITY 3) Shareholders' funds 4) Debit-to-equity ratio] from these 4 ratios, I checked the probability chances of direct and indirect relationships to the default or non-default. If my P-value is greater than 0.05, it’s a direct relationship to default, and if my P-value is less than 0.05, it’s an indirect relationship to non-default. Then, I decided to check my model's accuracy rate I used some libraries for SkLearn. I got 96% model accuracy.